feat: add permission administration API with CRUD operations for groups and nodes
- Introduced new permission routes for managing permission groups and nodes.
- Implemented handlers for creating, updating, and deleting permission groups.
- Added functionality to grant and revoke permission nodes for groups and players.
- Enhanced the OpenAPI documentation to include new permission endpoints.
- Updated existing models and requests to accommodate permission-related data.
- Added tests to ensure the correctness of the new permission routes and their error handling.

+++ internal/realm/catalog/commands/buy/command_test.go
@@ -114,10 +114,10 @@ type buyManager struct {
 }
 
 // Pages supplies unused catalog tree behavior.
-func (buyManager) Pages(context.Context, int32, bool) ([]catalogmodel.Page, error) { return nil, nil }
+func (buyManager) Pages(context.Context, int64, bool) ([]catalogmodel.Page, error) { return nil, nil }
 
 // Page returns the offered item.
-func (manager *buyManager) Page(context.Context, int64, int32, bool) (catalogmodel.Page, []catalogmodel.Item, error) {
+func (manager *buyManager) Page(context.Context, int64, int64, bool) (catalogmodel.Page, []catalogmodel.Item, error) {
 	if manager.pageItems != nil {
 		return catalogmodel.Page{}, manager.pageItems, nil
 	}
